requirements.txt
audioread==3.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
fonttools==4.47.2
idna==3.6
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
matplotlib==3.8.2
msgpack==1.0.7
numba==0.58.1
numpy==1.26.3
packaging==23.2
pandas==2.1.4
pillow==10.2.0
platformdirs==4.1.0
pooch==1.8.0
psycopg2==2.9.9
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
threadpoolctl==3.2.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
